#11 Use Java_Home #6 Code refactoring before adding jlink support
